requirements.txt
aiohttp==3.9.4
Django==3.2.25
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
esdk-obs-python==3.23.12
icalendar==4.1.0
mysqlclient==2.0.3
PyMySQL==1.1.1
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0
uwsgi==2.0.22
pycryptodome==3.20.0
yarl==1.9.4
httpx==0.26.0
bilibili-api-python==16.2.0
kafka-python==2.0.2
drf-yasg